+++ src/main/java/com/robrit/snad/blocks/SnadBlock.java
@@ -3,19 +3,22 @@ package com.robrit.snad.blocks;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
+import com.robrit.snad.Snad;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.server.level.ServerLevel;
-import net.minecraft.tags.FluidTags;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.FallingBlock;
 import net.minecraft.world.level.block.state.BlockState;
-import net.neoforged.neoforge.common.IPlantable;
-import net.neoforged.neoforge.common.PlantType;
+import net.minecraft.world.level.material.FluidState;
+import net.neoforged.neoforge.common.util.TriState;
 
 import javax.annotation.ParametersAreNonnullByDefault;
+import java.util.Iterator;
 
 public class SnadBlock extends FallingBlock {
     public static final MapCodec<SnadBlock> CODEC = RecordCodecBuilder.mapCodec(
@@ -38,22 +41,24 @@ public class SnadBlock extends FallingBlock {
 
     @Override
     @ParametersAreNonnullByDefault
-    public boolean canSustainPlant(BlockState state, BlockGetter world, BlockPos pos, Direction facing, IPlantable plantable) {
-        if (plantable.getPlantType(world, pos) == PlantType.DESERT) {
-            return true;
-        } else if (plantable.getPlantType(world, pos) == PlantType.BEACH) {
-            for (Direction direction : Direction.Plane.HORIZONTAL) {
-                boolean isWater = world.getFluidState(pos.relative(direction)).is(FluidTags.WATER);
-                boolean isFrostedIce = world.getBlockState(pos.relative(direction)).is(Blocks.FROSTED_ICE);
-                if (!isWater && !isFrostedIce) {
-                    continue;
+    public TriState canSustainPlant(BlockState state, BlockGetter level, BlockPos soilPos, Direction facing, BlockState plant) {
+        if (plant.is(Snad.SNAD_PLACEABLE_CROPS)) {
+            if (plant.is(Snad.SNAD_REQUIRES_WATER)) {
+                for (Direction direction : Direction.Plane.HORIZONTAL) {
+                    BlockState blockstate1 = level.getBlockState(soilPos.relative(direction));
+                    FluidState fluidstate = level.getFluidState(soilPos.relative(direction));
+                    if (state.canBeHydrated(level, soilPos.above(), fluidstate, soilPos.relative(direction)) || blockstate1.is(Blocks.FROSTED_ICE)) {
+                        return TriState.TRUE;
+                    }
                 }
 
-                return true;
+                return TriState.FALSE;
             }
+
+            return TriState.TRUE;
         }
 
-        return false;
+        return TriState.FALSE;
     }
 
     @Override